+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [3.0.8] - 2025-05-06
 ### Fixed
-- Fix potential bug with Weapon Container. (Related Issues: [#160], [#161])
+- [Foundry V13] Fix duplicated weapons set on switch. (Related Issues: [#161])
+
+## [3.0.8] - 2025-05-06
+### Fixed
+- Fix potential bug with Weapon Container. (Related Issues: [#160])
 
 ### Features
 - Add CTRL/ALT modifier on Initiative button. (Related Issue: [#158])